safari-600.7-branch/LayoutTests/fast/dom/shadow/tab-order-iframe-and-shadow-expected.txt

CONSOLE MESSAGE: line 104: Uncaught ReferenceError: log is not defined
This tests that pressing Tab key should traverse into iframe and shadow tree, and pressing Shift-Tab should reverse the order.

Should move from input-01 to input-13 in forward
PASS
Should move from input-13 to input-15 in forward
PASS
Should move from input-15 to input-02 in forward
PASS
Should move from input-02 to host-01/input-03 in forward
PASS
Should move from host-01/input-03 to input-04 in forward
PASS
Should move from input-04 to iframe/input-06 in forward
PASS
Should move from iframe/input-06 to iframe/host-02 in forward
PASS
Should move from iframe/host-02 to iframe/host-02/input-09 in forward
PASS
Should move from iframe/host-02/input-09 to iframe/host-02/input-08 in forward
PASS
Should move from iframe/host-02/input-08 to iframe/input-12 in forward
PASS
Should move from iframe/input-12 to iframe/input-11 in forward
PASS
Should move from iframe/input-11 to iframe/input-05 in forward
PASS
Should move from iframe/input-05 to input-14 in forward
PASS
Should move from input-14 to iframe/input-05 in backward
PASS
Should move from iframe/input-05 to iframe/input-11 in backward
PASS
Should move from iframe/input-11 to iframe/input-12 in backward
PASS
Should move from iframe/input-12 to iframe/host-02/input-08 in backward
PASS
Should move from iframe/host-02/input-08 to iframe/host-02/input-09 in backward
PASS
Should move from iframe/host-02/input-09 to iframe/host-02 in backward
PASS
Should move from iframe/host-02 to iframe/input-06 in backward
PASS
Should move from iframe/input-06 to input-04 in backward
PASS
Should move from input-04 to host-01/input-03 in backward
PASS
Should move from host-01/input-03 to input-02 in backward
PASS
Should move from input-02 to input-15 in backward
PASS
Should move from input-15 to input-13 in backward
PASS
Should move from input-13 to input-01 in backward
PASS
